requirements.txt
conda install pytorch==1.12.1 torchvision==0.13.1 cudatoolkit=11.3 -c pytorch
timm==0.6.12
pillow
tqdm
open_clip_torch
scipy